Encoders can fail, Codec's are GADTs.
Encoders can now return a Maybe Builder, allowing for sensible
prioritised choice between codecs, which is useful now that:

Codecs' are GADTs, there are two constructors: Plain, which is in
spirit the same that we had before, and Wrap, which is defined in
terms of Maybe-producing type-mangling functions and a smaller
codec.

There is also an operator for prioritised choice between codecs,
which handles encoding as well as decoding now.

This makes dealing with sum types much nicer: see the bencoding
test/example.